requirements.txt
# Virtual Pamudu Chat API Dependencies

# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Firebase
firebase-admin>=6.4.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20

# Data & Validation
pydantic>=2.5.0
python-dotenv>=1.0.0

# GitHub API
PyGithub>=2.1.1

# HTTP & Scraping
httpx>=0.26.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0

# YouTube (optional - for transcripts)
youtube-transcript-api>=0.6.2
